HiNet/requirements.txt
absl-py==0.11.0
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
blinker==1.4
brotlipy==0.7.0
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
cryptography==2.9.2
cycler==0.10.0
cytoolz==0.11.0
dask==2020.12.0
decorator==4.4.2
google-auth==1.24.0
google-auth-oauthlib==0.4.2
grpcio==1.31.0
idna==2.10
imageio==2.9.0
importlib-metadata==2.0.0
markdown==3.3.3
matplotlib==2.2.3
natsort==7.1.0
networkx==2.5
numpy==1.19.2
oauthlib==3.1.0
olefile==0.46
opencv-python==3.4.2.17
pandas==1.2.0
pillow==6.0.0
pip==20.3.3
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyjwt==2.0.0
pyopenssl==20.0.1
pyparsing==2.4.7
pyqt5==5.9.2
pysocks==1.7.1
python-dateutil==2.8.1
pytz==2020.4
pywavelets==1.1.1
pyyaml==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.14.2
scikit-learn==0.20.3
scipy==1.5.2
setuptools==51.0.0
sip==4.19.8
six==1.15.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.6.0
torch==1.0.1
torchvision==0.2.2
tornado==6.1
tqdm==4.54.1
typing-extensions==3.7.4.3
urllib3==1.26.2
werkzeug==1.0.1
wheel==0.36.2
yarl==1.5.1
zipp==3.4.0

# Additional packages from pip section in YAML file
backcall==0.2.0
coloredlogs==15.0
et-xmlfile==1.0.1
humanfriendly==9.1
ipdb==0.13.7
ipython==7.22.0
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.18.0
